wasmdeployqt: silence msvc warnings
* use size_t for qHash overload
* return false in copyPreloadQmlImports error path.

Amends b6cdc2a6ef89e72470eb24e7cf18eae4d861b20c

+++ src/tools/wasmdeployqt/common.h
@@ -18,7 +18,7 @@ struct PreloadEntry
     }
 };
 
-inline uint qHash(const PreloadEntry &key, uint seed = 0)
+inline size_t qHash(const PreloadEntry &key, size_t seed = 0)
 {
     return qHash(key.source, seed) ^ qHash(key.destination, seed);
 }
